backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
azure-cognitiveservices-speech==1.34.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
pydub==0.25.1
SpeechRecognition==3.10.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
